Get rid of type conversion warning.
git-svn-id: http://svn.eol.ucar.edu/svn/eol/common/trunk/pentek@7114 da89c675-a4d7-0310-8a15-9e7546503515

+++ p7142.cpp
@@ -163,8 +163,8 @@ p7142dn::read(char* buf, int bufsize) {
     int wl = 200;
     double fact = 1.0 + 0.05*(1.0*rand())/RAND_MAX;
     for (int i = 0; i < bufsize/2; i = i + 2) {
-      sbuf[i] = (10000.0 * sin(2.0*M_PI*i/wl)*fact);
-      sbuf[i+1] = (10000.0 * cos(2.0*M_PI*i/wl)*fact);
+      sbuf[i] = (short int)(10000.0 * sin(2.0*M_PI*i/wl)*fact);
+      sbuf[i+1] = (short int)(10000.0 * cos(2.0*M_PI*i/wl)*fact);
     }
     _bytesRead += bufsize;
